Add pool support for exec-zfs
This commit starts teaching exec-zfs how to parse `zpool` output. It
also makes it so the script is testable against a collection of known
outputs.

You can override the commands used to gather ZFS data, allowing
you to inject your own data to be parsed. For example, to force
exec-zfs.sh to parse the output of zpool on a mirrored pool:

```
$ LIST_DATASETS_CMD=" " \
    LIST_POOLS_CMD="echo test" \
    POOL_STATUS_CMD="cat tests/mirror-10-minute-scrub.status" \
    exec-zfs.sh
```
This will override the script so that no datasets are parsed, a
fake 'test' pool has its status read, and that the pool status to
be parsed is the output contained in the
`tests/mirror-10-minute-scrub.stats` file will be used instead.

 Please enter the commit message for your changes. Lines starting